backend/requirements.txt
flask
python-dotenv
cryptography
waitress
opencv-python-headless
numpy
